requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
aiosqlite==0.20.0
annotated-types==0.6.0
anyio==4.3.0
apparmor==3.0.4
APScheduler==3.10.4
aspose-cells==24.4.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.4
bs4==0.0.2
captcha==0.5.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cryptography==3.4.8
cycler==0.12.1
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.6.1
email_validator==2.1.1
et-xmlfile==1.1.0
fastapi==0.111.0
fastapi-cli==0.0.2
fonttools==4.51.0
frozenlist==1.4.1
h11==0.14.0
httpcore==1.0.5
httplib2==0.20.2
httptools==0.6.1
httpx==0.27.0
idna==3.7
importlib-metadata==4.6.4
iotop==0.6
jeepney==0.7.1
Jinja2==3.1.3
JPype1==1.5.0
keyring==23.5.0
kiwisolver==1.4.5
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
LibAppArmor==3.0.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
more-itertools==8.10.0
multidict==6.0.5
netifaces==0.11.0
numpy==1.26.4
oauthlib==3.2.0
openpyxl==3.1.2
orjson==3.10.3
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
pyTelegramBotAPI==4.17.0
python-apt==2.4.0+ubuntu3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==5.4.1
requests==2.31.0
rich==13.7.1
SecretStorage==3.3.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
soupsieve==2.5
starlette==0.37.2
telebot==0.0.5
typer==0.12.3
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
ujson==5.9.0
unattended-upgrades==0.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
wadllib==1.3.6
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
zipp==1.0.0
